Extracting GWT incompatible methods from Resource to Resources
